Remove Connection's dependency on EventHistory
This commit removes the dependency of module Connection on module
EventHistory. This became possible by making Connection output a file
change into a temporary file with an extension that distinguishes from
regular files and is ignored by ClientMonitor.

After outputting to the temporary file, control is passed over to
ServerMonitor, which is assigned to manipulate EventHistory, apply the
changes of the temporary file and finally delete it.