Add automatic reconnection system for the client
This commit changes how the client app connects to the service. Now, it
doesn't know beforehand the address of the servers, but must
communicate with a binding agent to get that information.

From now on, we assume that the primary server may fail. Thus, this
commit addresses automatic reconnection with the service by introducing a
reconnection system, consisting of a thread that executes a loop in
HeartbeatMonitor, which constantly checks the server's liveliness.

This commit makes the responsibility of threads that use Connection
to deal with failed message deliveries. However, no actual code that
deals with exceptions thrown by Connection was written so far. Such
issue will be dealt on following commits.