Make several bug fixes on the client
Prior to this commit, the client kept crashing after running into a
segmentation fault. It used to happen when the client connected to the
binding agent before any server had the opportunity to. Thus, the client
wouldn't create any new socket, leaving the internal server socket
descriptors in Connection with value -1. Then, when a thread tried to
read or write from/to the socket, it would cause a segmentation fault.